pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "noteflow-v2"
version = "2.0.0-alpha"
description = "A modular, reversible document processing pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Maxime Fournes" }
]
dependencies = [
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pyyaml>=6.0",
    "aiofiles>=23.2.0",
    "watchfiles>=0.21.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]
ai = [
    "ai-core",  # Your existing AI library
]

[project.scripts]
noteflow = "core.api.server:main"

[tool.hatch.build.targets.wheel]
packages = ["core", "plugins"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

